02-05-2016

+++ isasuk/calendar/views.py
@@ -2,16 +2,19 @@ from django.shortcuts import render_to_response
 from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
 from django.template import RequestContext
+from django.db import transaction
+
 from isasuk.meeting.models import Meeting
 from .forms import AddEventForm
 from ..meeting.models import Event
 import json
 import datetime
 
+@transaction.atomic
 @login_required
 def calendar_view(request):
     form = AddEventForm()
-    if 'add' in request.POST:
+    if request.user.details.role == 'member' and 'add' in request.POST:
       form = AddEventForm(request.POST)
       if form.is_valid():
         date = datetime.datetime.strptime(request.POST.get('date'), '%d/%m/%Y %H:%M').strftime('%Y-%m-%d %H:%M')
@@ -21,7 +24,7 @@ def calendar_view(request):
         )
         event.save()
         form = AddEventForm()
-    elif 'delete' in request.POST:
+    elif request.user.details.role == 'member' and 'delete' in request.POST:
       event = Event.objects.get(id=request.POST.get('id'))
       event.delete()
     events = Event.objects.all().filter(date__gt=datetime.datetime.now()).order_by('date')
@@ -34,14 +37,8 @@ def calendar_view(request):
         context_instance=RequestContext(request),
       )
 
+@transaction.atomic
 @login_required
-def add_event(request):
-    return HttpResponse(json.dumps(events))
-
-@login_required
-def remove_event(request):
-    return HttpResponse(json.dumps(events))
-
 def events_view(request):
     meetings = Meeting.objects.all()
     e = Event.objects.all()#.filter(date__gt=datetime.datetime.now())
@@ -49,12 +46,12 @@ def events_view(request):
     for meeting in meetings:
       events.append({
         'title': meeting.title,
-        'start': meeting.date.isoformat()
+        'start': meeting.date.isoformat(),
       })
     for event in e:
       events.append({
         'title': event.title,
-        'start': event.date.isoformat()
+        'start': event.date.isoformat(),
       })
     return HttpResponse(json.dumps(events))
